+++ src/routes/comments-router.ts
@@ -26,12 +26,19 @@ commentsRouter.put('/:id',
     contentValidation,
     inputValidationMw,
     async (req: Request, res: Response) => {
-        let updateStatus: boolean = await commentService.updateComment(req.params.id, req.body.content)
-        if (updateStatus) {
-            res.sendStatus(STATUSES_HTTP.NO_CONTENT_204)
-        } else {
+        let foundComment: CommentViewModel | null = await commentService.findCommentById(req.params.id)
+        if (!foundComment) {
             res.sendStatus(STATUSES_HTTP.NOT_FOUND_404)
+            return;
         }
+
+        if(foundComment.commentatorInfo.userId !== req.user!.id) {
+            res.sendStatus(STATUSES_HTTP.FORBIDDEN_403)
+            return;
+        }
+
+        let updateStatus: boolean = await commentService.updateComment(req.params.id, req.body.content)
+        res.sendStatus(STATUSES_HTTP.NO_CONTENT_204)
     }
 )
 
@@ -39,11 +46,19 @@ commentsRouter.delete('/:id',
     authenticationCheckBearer,
     inputValidationMw,
     async (req: Request, res: Response) => {
-        let deletionStatus: boolean = await commentService.deleteComment(req.params.id)
-        if (deletionStatus) {
-            res.sendStatus(STATUSES_HTTP.NO_CONTENT_204)
-        } else {
+
+        let foundComment: CommentViewModel | null = await commentService.findCommentById(req.params.id)
+        if (!foundComment) {
             res.sendStatus(STATUSES_HTTP.NOT_FOUND_404)
+            return;
         }
+
+        if(foundComment.commentatorInfo.userId !== req.user!.id) {
+            res.sendStatus(STATUSES_HTTP.FORBIDDEN_403)
+            return;
+        }
+
+        let deletionStatus: boolean = await commentService.deleteComment(req.params.id)
+        res.sendStatus(STATUSES_HTTP.NO_CONTENT_204)
     }
 )
